--- py-opsml/python/opsml/scouter/logging/__init__.py
@@ -1,15 +1,0 @@
-# type: ignore
-# pylint: disable=no-name-in-module
-from .. import scouter  # noqa: F401
-
-LogLevel = scouter.logging.LogLevel
-WriteLevel = scouter.logging.WriteLevel
-LoggingConfig = scouter.logging.LoggingConfig
-RustyLogger = scouter.logging.RustyLogger
-
-__all__ = [
-    "LogLevel",
-    "WriteLevel",
-    "LoggingConfig",
-    "RustyLogger",
-]
