+++ py-opsml/python/opsml/scouter/types/__init__.py
@@ -2,10 +2,10 @@
 # pylint: disable=no-name-in-module
 from .. import scouter
 
-CommonCrons = scouter.types.CommonCrons
-DataType = scouter.types.DataType
-DriftType = scouter.types.DriftType
-ScouterError = scouter.types.ScouterError
+CommonCrons = scouter._types.CommonCrons
+DataType = scouter._types.DataType
+DriftType = scouter._types.DriftType
+ScouterError = scouter._types.ScouterError
 
 __all__ = [
     "DriftType",
