+++ py-opsml/src/lib.rs
@@ -1,6 +1,5 @@
 pub mod card;
 pub mod cli;
-pub mod core;
 pub mod data;
 pub mod error;
 pub mod experiment;
@@ -20,8 +19,6 @@ fn opsml(m: &Bound<'_, PyModule>) -> PyResult<()> {
     m.add_wrapped(wrap_pymodule!(scouter::scouter))?;
     m.add_wrapped(wrap_pymodule!(potato_head::potato_head))?;
 
-    // core
-    m.add_wrapped(wrap_pymodule!(core::core))?;
     m.add_wrapped(wrap_pymodule!(data::data))?;
     m.add_wrapped(wrap_pymodule!(model::model))?;
     m.add_wrapped(wrap_pymodule!(card::card))?;
