+++ py-opsml/tests/conftest.py
@@ -1,6 +1,7 @@
 import pytest
 
-from opsml.core import Feature, RustyLogger, LoggingConfig, LogLevel
+from opsml.model import Feature
+from opsml.logging import RustyLogger, LoggingConfig, LogLevel
 from opsml.card import RegistryTestHelper
 from opsml.potato_head import Prompt
 from typing import Tuple, Dict
