+++ py-opsml/tests/interfaces/conftest.py
@@ -6,7 +6,7 @@ import numpy as np
 from numpy.typing import NDArray
 from datetime import datetime, timedelta
 from opsml.data import DataInterface
-from opsml.core import SaveName, Suffix, SaverPath
+from opsml.types import SaveName, Suffix, SaverPath
 from pathlib import Path
 import joblib  # type: ignore
 from typing import Dict
